Fix PC system type mismatches for 32-bit registers
- Updated PC CPU wrapper to handle u32 registers
- Fixed all I/O port operations to cast between u32/u16
- Updated CpuRegisters and DebugInfo structs to use u32
- Fixed GUI debug overlay to display 32-bit values
- Fixed most test assertions to compare with u32 types
- 232 out of 237 PC tests passing (5 tests need investigation)

+++ crates/systems/pc/src/lib.rs
@@ -345,16 +345,16 @@ impl PcSystem {
 #[derive(Debug, Clone)]
 pub struct DebugInfo {
     pub cs: u16,
-    pub ip: u16,
-    pub ax: u16,
-    pub bx: u16,
-    pub cx: u16,
-    pub dx: u16,
-    pub sp: u16,
-    pub bp: u16,
-    pub si: u16,
-    pub di: u16,
-    pub flags: u16,
+    pub ip: u32,
+    pub ax: u32,
+    pub bx: u32,
+    pub cx: u32,
+    pub dx: u32,
+    pub sp: u32,
+    pub bp: u32,
+    pub si: u32,
+    pub di: u32,
+    pub flags: u32,
     pub cycles: u64,
 }
